oss/requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
Django==3.2.7
firebase-admin==5.0.2
google-api-core==1.31.2
google-api-python-client==2.20.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.1
google-cloud-storage==1.42.1
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.40.0
httplib2==0.19.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jwt==1.2.0
msgpack==1.0.2
mysqlclient==2.0.3
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.0
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
qrcode==7.3
razorpay==1.2.0
requests==2.26.0
rsa==4.7.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
zope.interface==5.4.0
